Initial version of python frags to features code.
This code incorporates the essentials of the old Perl algorithm, but
instead constructs an object-oriented data model. A strong effort is
made to avoid FeatureType-dependent code, and to sequester any such code
that is included into callbacks. This process is incomplete, and some of
the final FeatureType-dependent steps of the old Perl code have not been
implemented.