Database access optimizations and documentation.
Rewrote the frags_to_features functions to only access the database once
for each feature type, at the very beginning of the algorithm. The data
pulled from the database is stored in internal object for use by the
algorithm.

Also, modified the transaction middleware for blat() to prevent
unnecessary commit commands: instead of committing after every
seq_feature.save(), it now commits at the end of all of the saves().
This drastically reduced the runtime of database addition.

Finally, comment strings and comments have been updated for most
relevant blocks of code. Some of the comments are redundant, as the
function names are clear, but they help with the structure of the file.