Parsing of features from JSON abstracted out.
Major restructuring of the API. Instead of a single closure function,
the API now consits of one major outside scope function, GiraffeDraw(),
which takes no arguments and returns an object with several methods. The
object is used to hold the parsed in feature data.

The object contains a read() method, which is called on a JSON list of
feature data. The method takes that data and stores it in a global list
of basic Feature objects. These basic Feature objects mostly contain the
data members and methods that are responsible for the raw data of the
feature class: names, types, enzyme cut counts, etc.

The object also contains draw_<foo>_map methods. Currently, only the
circular_map method is implemented, with space left for the linear_map
method. These methods take the basic feature list and extend the basic
Feature objects into <Foo>Feature objects, which contain the additional
data and methods necessary to make the drawing work.

Code comments were updated to reflect this change.