Gene variant name modification.
I finally added back in the last bit of feature-dependant code from the
old perl script. This was the part that, for genes, calculates a few
extra properties: whether it's a variant, whether it's a gap, or whether
it's a subset. Currently, there is a feature-dependant part of the
to_sequence_feature() method that calculates this extra data and stores
it in the Sequence_Feature model (which has been extended with new
properties to accommodate this). Then, the to_dict() method of
Sequence_Feature uses the extra fields to modify the "feature" entry
of the dictionary.

These additions may slow down or compliate database access, so I was
hesitant to add them (to say nothing of the feature-dependence), but
this is previously-available functionality, so, unless it causes
problems, it should still be made available, one way or another.