Regroup linear labels
Now, the 0th and 2nd lists are all on top, left-justified, and the 1st and 3rd are all on the bottom, right-justified. This lets each list be offset far to one side from the center of the section it is responsible for. The offset allows all the label lines to point in the same direction (feature: left, label:right on top, and vice versa on bottom), and makes it easier to avoid crossing label lines. Parameters may still need to be adjusted to make this work better.